making layout file for courses reusable
