+++ 3sp/src/app/api/auth/[...nextauth]/route.ts
@@ -39,6 +39,7 @@ export const authOptions: NextAuthOptions = {
     },
   },
   secret: process.env.NEXTAUTH_SECRET,
+  debug: true,
 };
 
 const handler = NextAuth(authOptions);
